move external storage to separate file
